further outsources foods.js and foodCreate.js code to our HTMLHelper class
