+++ lib/foodAjax.js
@@ -1,5 +1,6 @@
 var $ = require('jQuery');
 var foods = require('./foods.js');
+var HTMLHelper = require('./htmlHelper.js')
 
 $(document).ready(function(){
   $.ajax({
@@ -19,7 +20,7 @@ function postAjax(formattedFood) {
     data: formattedFood,
     dataType: "json",
     success: function addFoodFunction(data){
-      foods.addNewFoodToTable(data);
+      HTMLHelper.addNewFoodToTable(data);
     }
   });
 };
