+++ lib/foods.js
@@ -24,20 +24,6 @@ function createFoodsTable(unsortedFoods){
   return HTMLHelper.createTableTemplate(foodsArray, '.food-table')
 }
 
-function foodString(food) {
-  return "<tr class='tbl-row'><td class='cell-one food-cell' contenteditable='true'>"
-  + "<span class='replaceme-name' contenteditable='true'>"
-  + food.name
-  + "</span>"
-  + "</td><td class='cell-two cal-cell'>"
-  + "<span class='replaceme-cal' contenteditable='true'>"
-  + food.calories
-  + "</span>"
-  + "</td><td class='delete-cell'><button class='delete-food-btn'><img class='delete-food-img' src='lib/images/delete_red.png'/></button></td><td class='food-id' style='visibility:hidden;'>"
-  + food.id
-  + "</td></tr>"
-}
-
 function sortFoodsArray(unsortedFoods) {
   sortedFoods = unsortedFoods.sort(function(a, b) {
     return (b.id) - (a.id);
@@ -45,8 +31,6 @@ function sortFoodsArray(unsortedFoods) {
   return sortedFoods
 }
 
-function addNewFoodToTable(food) {
-  $(".food-table tr:first").after(foodString(food));
-}
 
-module.exports = {logErrors, makeFoods, createFoodsTable, addNewFoodToTable}
+
+module.exports = {logErrors, makeFoods, createFoodsTable}
