+++ lib/htmlHelper.js
@@ -31,6 +31,22 @@ class HTMLHelper {
     $('#food-calories').focus();
     focusSet = true;
   }
-}
+  static addNewFoodToTable(food) {
+    $(".food-table tr:first").after(foodString(food));
+    function foodString(food) {
+      return "<tr class='tbl-row'><td class='cell-one food-cell' contenteditable='true'>"
+      + "<span class='replaceme-name' contenteditable='true'>"
+      + food.name
+      + "</span>"
+      + "</td><td class='cell-two cal-cell'>"
+      + "<span class='replaceme-cal' contenteditable='true'>"
+      + food.calories
+      + "</span>"
+      + "</td><td class='delete-cell'><button class='delete-food-btn'><img class='delete-food-img' src='lib/images/delete_red.png'/></button></td><td class='food-id' style='visibility:hidden;'>"
+      + food.id
+      + "</td></tr>"
+    }
+  }
+};
 
 module.exports = HTMLHelper
