+++ lib/index.js
@@ -1,5 +1,6 @@
 require('./food.js')
 require('./foods.js')
+require('./htmlHelper.js')
 require('./foodAjax.js')
 require('./foodCreate.js')
 require('./foodDelete.js')
@@ -7,4 +8,3 @@ require('./foodEdit.js')
 require('./foodSearch.js')
 require('./indexMealTables.js')
 require('./indexMealFoodDelete.js')
-require('./htmlHelper.js')
